refactor: 함수명에 맞게 compareInputNum으로 모듈명 변경
