+++ src/modules/compareInputNum.js
