+++ src/modules/game.js
@@ -1,6 +1,6 @@
 const MissionUtils = require('@woowacourse/mission-utils');
 const { isThreeRandomNumbers, isNotZeroInInputAndNotDoubleNum } = require('./valid');
-const { compareInputNum } = require('./compareNums');
+const { compareInputNum } = require('./compareInputNum');
 const { generateThreeRandomNums } = require('./generateNums');
 
 function gameFirstStart() {
@@ -15,11 +15,11 @@ function baseBallGame(computerNums) {
     MissionUtils.Console.print(compareInputNum(computerNums, userInputNums));
     if (computerNums !== userInputNums) return baseBallGame(computerNums);
     MissionUtils.Console.print('3개의 숫자를 모두 맞히셨습니다! 게임 종료');
-    MissionUtils.Console.readLine('게임을 새로 시작하려면 1, 종료하려면 2를 입력하세요.\n', (userInputNums) => {
-      if (!(userInputNums === '1' || userInputNums === '2')) throw '잘못된 입력입니다.';
-      if (userInputNums === '1') return baseBallGame(generateThreeRandomNums());
-      MissionUtils.Console.close();
-    });
+  });
+  MissionUtils.Console.readLine('게임을 새로 시작하려면 1, 종료하려면 2를 입력하세요.\n', (userInputNums) => {
+    if (!(userInputNums === '1' || userInputNums === '2')) throw '잘못된 입력입니다.';
+    if (userInputNums === '1') return baseBallGame(generateThreeRandomNums());
+    MissionUtils.Console.close();
   });
   return;
 }
